Fix NPE when using menu bar on Linux GTK3
When running on Linux with GTK3, WB attempts to draw the popup menu
using the handle of containing menu. Because this menu is not rendered
properly, the handle returns null, leading to a NullPointerException.

To work around this problem, a mockup of the popup is created instead,
similarly to how it's done on MacOSX with Cocoa. The code has been moved
from the MacOSX bundle to the SWT bundle, so that it can be used for
both Linux and Mac.
