+++ org.eclipse.wb.os.linux/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Require-Bundle: org.eclipse.swt,
  org.eclipse.jface,
  org.eclipse.ui,
  org.eclipse.wb.core;bundle-version="1.12.0",
+ org.eclipse.wb.swt,
  com.google.guava;bundle-version="31.1.0"
 Bundle-Activator: org.eclipse.wb.internal.os.linux.Activator
 Bundle-ActivationPolicy: lazy
