+++ org.eclipse.wb.swt/src/org/eclipse/wb/internal/swt/check.png
